c2-platform/pyproject.toml
[project]
name = "c2-platform"
version = "0.1.0"
description = "Lattice-style C2 platform with web UI and subsystem command API"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.26.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = []
